requirements.txt
requests
tqdm
mosspy
pyyaml
addict